requirements.txt
pdfplumber>=0.10.0
ocrmypdf>=16.0.0
openai>=1.0.0
pyyaml>=6.0
docxtpl>=0.18.0
Flask>=3.0
mammoth>=1.6
beautifulsoup4>=4.12
